Issues Resolving
Timeline which we initially made 2nd one that is perfect where on
tapping the image from timeline the image was opening in zoom in mode.
(made it as it was previoulsy)

Tag a game-add a shot

 Home screen (todays gameà game selectionà enter the field screen
buttonà list of 2 teams to select which to tag and add to <team name>
shot. Next time when user comes to this same game and taps on enter the
field he will go to the game timeline posts.

Change trending post text to Famed Page

Tapping top right + icon will show a screen showing my Buddies

“Buddies at the game” under enter the field screen

Add profile picture to timeline

When add the Clippers it's adding Stanford too. I didn't add Stanford

Menu navigation cutting off